requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.27
Werkzeug==2.2.3
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
itsdangerous==2.1.2
python-dotenv==1.0.0
google-generativeai==0.8.5